refactor: separate WorkspaceMessageProcessor
